refactoring repository services to remove dependency on core services
