+++ legend-depot-artifacts-purge/src/test/java/org/finos/legend/depot/store/artifacts/services/TestArtifactsPurgeService.java
@@ -17,12 +17,12 @@ package org.finos.legend.depot.store.artifacts.services;
 
 import org.finos.legend.depot.artifacts.repository.api.ArtifactRepository;
 import org.finos.legend.depot.artifacts.repository.domain.ArtifactType;
-import org.finos.legend.depot.artifacts.repository.domain.VersionMismatch;
-import org.finos.legend.depot.artifacts.repository.services.RepositoryServices;
+import org.finos.legend.depot.domain.version.VersionMismatch;
 import org.finos.legend.depot.domain.DatesHandler;
 import org.finos.legend.depot.domain.api.MetadataEventResponse;
 import org.finos.legend.depot.domain.project.StoreProjectData;
 import org.finos.legend.depot.domain.project.StoreProjectVersionData;
+import org.finos.legend.depot.services.VersionsMismatchService;
 import org.finos.legend.depot.services.api.entities.ManageEntitiesService;
 import org.finos.legend.depot.services.api.projects.ManageProjectsService;
 import org.finos.legend.depot.services.entities.ManageEntitiesServiceImpl;
@@ -82,9 +82,9 @@ public class TestArtifactsPurgeService extends TestStoreMongo
     protected UpdateEntities entitiesStore = new EntitiesMongo(mongoProvider);
     protected UpdateFileGenerations fileGenerationsStore = new FileGenerationsMongo(mongoProvider);
     protected ArtifactRepository repository = mock(ArtifactRepository.class);
-    protected RepositoryServices repositoryServices = new RepositoryServices(repository, projectsService);
+    protected VersionsMismatchService versionsMismatchService = new VersionsMismatchService(repository, projectsService);
     protected ManageEntitiesService entitiesService = new ManageEntitiesServiceImpl(entitiesStore, projectsService);
-    protected ArtifactsPurgeService purgeService = new ArtifactsPurgeServiceImpl(projectsService, repositoryServices, metricHandler);
+    protected ArtifactsPurgeService purgeService = new ArtifactsPurgeServiceImpl(projectsService, versionsMismatchService, metricHandler);
 
 
 
@@ -225,7 +225,7 @@ public class TestArtifactsPurgeService extends TestStoreMongo
         List<String> versions = projectsService.getVersions(TEST_GROUP_ID,TEST_ARTIFACT_ID);
         Assert.assertEquals(versions.size(), 3);
         Assert.assertEquals(3, fileGenerationsStore.getAll().size());
-        when(repositoryServices.findVersionsMismatches()).thenReturn(Collections.singletonList(versionMismatch));
+        when(versionsMismatchService.findVersionsMismatches()).thenReturn(Collections.singletonList(versionMismatch));
         //deleting the version not present in the repository
         purgeService.deprecateVersionsNotInRepository();
         versions = projectsService.getVersions(TEST_GROUP_ID,TEST_ARTIFACT_ID);
