+++ legend-depot-core-services/src/test/java/org/finos/legend/depot/services/TestVersionsMismatchService.java
@@ -13,13 +13,13 @@
 //  limitations under the License.
 //
 
-package org.finos.legend.depot.artifacts.repository.services;
+package org.finos.legend.depot.services;
 
 import org.finos.legend.depot.artifacts.repository.api.ArtifactRepository;
 import org.finos.legend.depot.artifacts.repository.api.ArtifactRepositoryException;
-import org.finos.legend.depot.artifacts.repository.domain.VersionMismatch;
 import org.finos.legend.depot.domain.project.StoreProjectData;
 import org.finos.legend.depot.domain.project.StoreProjectVersionData;
+import org.finos.legend.depot.domain.version.VersionMismatch;
 import org.finos.legend.depot.services.api.projects.ProjectsService;
 import org.finos.legend.sdlc.domain.model.version.VersionId;
 import org.junit.Assert;
@@ -34,11 +34,11 @@ import java.util.List;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class TestVersionsMismatch
+public class TestVersionsMismatchService
 {
     protected ArtifactRepository repository = mock(ArtifactRepository.class);
     protected ProjectsService projects = mock(ProjectsService.class);
-    protected RepositoryServices repositoryServices = new RepositoryServices(repository,projects);
+    protected VersionsMismatchService repositoryServices = new VersionsMismatchService(repository, projects);
 
     @Before
     public void setup() throws ArtifactRepositoryException
