+++ legend-depot-model/src/main/java/org/finos/legend/depot/domain/version/VersionMismatch.java
@@ -13,7 +13,7 @@
 //  limitations under the License.
 //
 
-package org.finos.legend.depot.artifacts.repository.domain;
+package org.finos.legend.depot.domain.version;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
